backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "backend for inventory management system"
authors = ["jakubpluta <plutakuba@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
uvicorn = "^0.18.2"
pydantic = "^1.9.1"
passlib = "^1.7.4"
tortoise-orm = "^0.19.2"
fastapi-mail = "^1.0.9"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
